build: rename module
