+++ .github/workflows/sphinx_docs.yml
@@ -1,6 +1,8 @@
 name: Docs
 
-on: push
+on:
+  push:
+    branches: [ "main" ]
 
 jobs:
   build:
