+++ docs/source/api.rst
@@ -4,7 +4,13 @@ API Reference
 .. autosummary::
    :toctree: generated
 
-.. automodule:: py_currency
-    :members:
-    :inherited-members:
-    :special-members: __init__
+.. automodule:: monepy
+   :members:
+   :inherited-members:
+   :special-members: __init__
+
+.. toctree::
+   :maxdepth: 2
+
+   submodules/currency
+   submodules/utils
