+++ docs/source/submodules/currency.rst
@@ -1,0 +1,11 @@
+Currency
+========
+
+.. autosummary::
+   :toctree: generated
+      :maxdepth: 2
+
+.. automodule:: monepy.currency
+   :members:
+   :inherited-members:
+   :special-members: __init__
