+++ docs/source/submodules/utils.rst
@@ -1,0 +1,10 @@
+Utils
+=====
+
+.. autosummary::
+   :toctree: generated
+      :maxdepth: 2
+
+.. automodule:: monepy.utils
+   :members:
+   :inherited-members:
