+++ src/monepy/__init__.py
@@ -1,0 +1,6 @@
+"""Monepy
+   ======
+   Currency classes to work with monetary values
+"""
+
+__version__ = "0.1"
