+++ src/monepy/currency/__init__.py
@@ -1,0 +1,6 @@
+"""Module with currency classes"""
+
+from .currencies import BRL, EUR, USD, JPY
+
+
+__all__ = ["BRL", "EUR", "USD", "JPY"]
