--- src/py_currency/__init__.py
@@ -1,5 +1,0 @@
-from .currencies import BRL, EUR, USD, JPY
-from .utils import sum_
-
-
-__all__ = ["BRL", "EUR", "USD", "JPY", "sum_"]
